Move twitter classes to a twitter package
