+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveConnectorFactory.java
@@ -29,6 +29,9 @@ import com.facebook.presto.spi.classloader.ClassLoaderSafeConnectorSplitManager;
 import com.facebook.presto.spi.classloader.ThreadContextClassLoader;
 import com.facebook.presto.spi.security.ConnectorAccessControl;
 import com.facebook.presto.spi.type.TypeManager;
+import com.facebook.presto.twitter.hive.MetastoreStaticClusterModule;
+import com.facebook.presto.twitter.hive.MetastoreZkDiscoveryBasedModule;
+import com.facebook.presto.twitter.hive.ZookeeperServersetMetastoreConfig;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableSet;
 import com.google.inject.Injector;
