+++ presto-hive/src/main/java/com/facebook/presto/twitter/hive/ZookeeperServersetHiveCluster.java
@@ -11,8 +11,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.twitter.hive;
 
+import com.facebook.presto.hive.HiveCluster;
+import com.facebook.presto.hive.HiveMetastoreClientFactory;
 import com.facebook.presto.hive.metastore.HiveMetastoreClient;
 import com.facebook.presto.spi.PrestoException;
 import com.google.common.net.HostAndPort;
