+++ presto-hive/src/main/java/com/facebook/presto/twitter/hive/util/UgiUtils.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive.util;
+package com.facebook.presto.twitter.hive.util;
 
 import org.apache.hadoop.security.UserGroupInformation;
